fix: fix int (#7082)

+++ master/internal/db/postgres_test_utils.go
@@ -328,9 +328,9 @@ func (db *PgDB) MustExec(t *testing.T, sql string, args ...any) sql.Result {
 
 // MockWorkspaces creates as many new workspaces as in workspaceNames and
 // returns a list of workspaceIDs.
-func MockWorkspaces(workspaceNames []string, userID model.UserID) ([]int, error) {
+func MockWorkspaces(workspaceNames []string, userID model.UserID) ([]int32, error) {
 	ctx := context.Background()
-	var workspaceIDs []int
+	var workspaceIDs []int32
 	var workspaces []model.Workspace
 
 	for _, workspaceName := range workspaceNames {
@@ -354,14 +354,14 @@ func MockWorkspaces(workspaceNames []string, userID model.UserID) ([]int, error)
 	}
 
 	for _, workspace := range workspaces {
-		workspaceIDs = append(workspaceIDs, workspace.ID)
+		workspaceIDs = append(workspaceIDs, int32(workspace.ID))
 	}
 
 	return workspaceIDs, nil
 }
 
 // CleanupMockWorkspace removes the specified workspaceIDs from the workspaces table.
-func CleanupMockWorkspace(workspaceIDs []int) error {
+func CleanupMockWorkspace(workspaceIDs []int32) error {
 	var workspaces []model.Workspace
 	_, err := Bun().NewDelete().Model(&workspaces).
 		Where("id IN (?)", bun.In(workspaceIDs)).
